Revert "Opening Balance entry screen completed. Hibernate to JPA mappings."
This reverts commit 5cd1411f06b80e998988811f4ef41980544b45c8.

+++ egov/egov-collectionweb/src/main/java/org/egov/collection/web/actions/receipts/AjaxBankRemittanceAction.java
@@ -109,7 +109,7 @@ public class AjaxBankRemittanceAction extends BaseFormAction {
         for (int i = 0; i < queryResults.size(); i++) {
             final Object[] arrayObjectInitialIndex = queryResults.get(i);
             final Bankbranch newBankbranch = new Bankbranch();
-            newBankbranch.setId(Integer.valueOf(arrayObjectInitialIndex[0].toString()).longValue());
+            newBankbranch.setId(Integer.valueOf(arrayObjectInitialIndex[0].toString()));
             newBankbranch.setBranchname(arrayObjectInitialIndex[1].toString());
             bankBranchArrayList.add(newBankbranch);
         }
